--- app/src/main/java/com/nadosunbae_andorid/presentation/ui/classroom/adapter/ClassRoomSeniorAdapter.kt
@@ -1,31 +1,0 @@
-package com.nadosunbae_andorid.presentation.ui.classroom.adapter
-
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-
-class ClassRoomSeniorAdapter : RecyclerView.Adapter<ClassRoomSeniorAdapter.ClassRoomSeniorViewHolder>() {
-
-
-    override fun onCreateViewHolder(
-        parent: ViewGroup,
-        viewType: Int
-    ): ClassRoomSeniorAdapter.ClassRoomSeniorViewHolder {
-
-    }
-
-    override fun onBindViewHolder(
-        holder: ClassRoomSeniorAdapter.ClassRoomSeniorViewHolder,
-        position: Int
-    ) {
-
-    }
-
-    override fun getItemCount(): Int {
-
-    }
-
-    inner class ClassRoomSeniorViewHolder(){
-
-
-    }
-}
