+++ app/src/main/res/layout/item_question_senior_off_question.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content">
@@ -13,7 +14,7 @@
         app:layout_constraintBottom_toBottomOf="parent">
 
         <ImageView
-            android:id="@+id/img_senior_title"
+            android:id="@+id/img_senior_title_off"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="12dp"
@@ -25,7 +26,7 @@
             android:layout_marginEnd="12dp"/>
 
         <TextView
-            android:id="@+id/text_senior_nickname"
+            android:id="@+id/text_senior_nickname_off"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/review_example_nickname"
@@ -33,13 +34,13 @@
             android:textSize="14sp"
             android:textColor="@color/black"
             android:includeFontPadding="false"
-            app:layout_constraintTop_toBottomOf="@+id/img_senior_title"
+            app:layout_constraintTop_toBottomOf="@+id/img_senior_title_off"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             android:layout_marginTop="4dp"/>
 
         <TextView
-            android:id="@+id/text_senior_major"
+            android:id="@+id/text_senior_major_off"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/start_major"
@@ -47,12 +48,12 @@
             android:textSize="12sp"
             android:textColor="@color/main_default"
             android:includeFontPadding="false"
-            app:layout_constraintTop_toBottomOf="@+id/text_senior_nickname"
+            app:layout_constraintTop_toBottomOf="@+id/text_senior_nickname_off"
             app:layout_constraintStart_toStartOf="parent"
             android:layout_marginStart="17dp"/>
 
         <TextView
-            android:id="@+id/text_senior_major_start"
+            android:id="@+id/text_senior_major_start_off"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/major_start"
@@ -60,8 +61,8 @@
             android:textSize="12sp"
             android:textColor="@color/gray_4"
             android:includeFontPadding="false"
-            app:layout_constraintTop_toTopOf="@+id/text_senior_major"
-            app:layout_constraintStart_toEndOf="@+id/text_senior_major"
+            app:layout_constraintTop_toTopOf="@+id/text_senior_major_off"
+            app:layout_constraintStart_toEndOf="@+id/text_senior_major_off"
             android:layout_marginStart="4dp"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
